Move around intercept property code to JqwikMicronautExtension
